+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added example to `terra-modal-manager` to show how to consume the new closeOnOutsideClick property.
+  
 * Changed
   * Updated the `terra-date-time-picker` example for field label.